refactor: modal overlay
