--- packages/react/src/components/dialog/index.ts
@@ -1,14 +1,0 @@
-export { default as Modal } from './components/Modal';
-export * from './components/Modal';
-
-export { default as ModalContent } from './components/DialogContent';
-export * from './components/DialogContent';
-
-export { default as ModalFooter } from './components/DialogFooter';
-export * from './components/DialogFooter';
-
-export { default as ModalHeader } from './components/DialogHeader';
-export * from './components/DialogHeader';
-
-export { default as ModalHeaderIcon } from './components/ModalHeaderIcon';
-export * from './components/ModalHeaderIcon';
