+++ packages/react/src/components/index.ts
@@ -6,7 +6,7 @@ export * from './button';
 export * from './button-group';
 export * from './card';
 export * from './collapsible';
-export * from './dialog';
+export * from './modal';
 export * from './dropdown';
 export * from './dropzone';
 export * from './empty-state';
@@ -17,6 +17,7 @@ export * from './loading';
 export * from './message';
 export * from './navigation';
 export * from './notification';
+export * from './overlay';
 export * from './page-header';
 export * from './pagination';
 export * from './panel';
