+++ packages/react/src/components/modal/components/ModalContent.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { Box } from '../../../layout';
 
-export interface DialogContentProps {
+export interface ModalContentProps {
   className?: string;
   style?: React.CSSProperties;
 }
 
-const ModalContent: React.FC<DialogContentProps> = ({ className, style, children }) => {
+const ModalContent: React.FC<ModalContentProps> = ({ className, style, children }) => {
   return (
     <Box className={className} style={style} display="flex" flexDirection="column" flex="1 1 auto" px="lg" py="md">
       {children}
@@ -15,8 +15,5 @@ const ModalContent: React.FC<DialogContentProps> = ({ className, style, children
 };
 
 ModalContent.displayName = 'ModalContent';
-
-/** @deprecated - use `ModalContent` instead */
-export const DialogContent = ModalContent;
 
 export default ModalContent;
