+++ packages/react/src/components/modal/components/ModalFooter.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { Box } from '../../../layout';
 
-export interface DialogFooterProps {
+export interface ModalFooterProps {
   className?: string;
   style?: React.CSSProperties;
 }
 
-const ModalFooter: React.FC<DialogFooterProps> = ({ className, style, children }) => {
+const ModalFooter: React.FC<ModalFooterProps> = ({ className, style, children }) => {
   return (
     <Box className={className} style={style} p="lg">
       {children}
@@ -15,8 +15,5 @@ const ModalFooter: React.FC<DialogFooterProps> = ({ className, style, children }
 };
 
 ModalFooter.displayName = 'ModalFooter';
-
-/** @deprecated - use `ModalFooter` instead */
-export const DialogFooter = ModalFooter;
 
 export default ModalFooter;
