+++ packages/react/src/components/modal/components/ModalHeader.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { Box, BoxProps } from '../../../layout';
 
-export interface DialogHeaderProps extends BoxProps {
+export interface ModalHeaderProps extends BoxProps {
   className?: string;
   style?: React.CSSProperties;
 }
 
-const ModalHeader: React.FC<DialogHeaderProps> = ({ className, style, children, ...rest }) => {
+const ModalHeader: React.FC<ModalHeaderProps> = ({ className, style, children, ...rest }) => {
   return (
     <Box className={className} style={style} display="flex" flexDirection="column" p="lg" {...rest}>
       {children}
@@ -15,8 +15,5 @@ const ModalHeader: React.FC<DialogHeaderProps> = ({ className, style, children, 
 };
 
 ModalHeader.displayName = 'ModalHeader';
-
-/** @deprecated - use `ModalHeader` instead */
-export const DialogHeader = ModalHeader;
 
 export default ModalHeader;
