+++ packages/react/src/components/modal/components/ModalHeaderIcon.tsx
