+++ packages/react/src/components/modal/index.ts
@@ -1,0 +1,14 @@
+export { default as Modal } from './components/Modal';
+export * from './components/Modal';
+
+export { default as ModalContent } from './components/ModalContent';
+export * from './components/ModalContent';
+
+export { default as ModalFooter } from './components/ModalFooter';
+export * from './components/ModalFooter';
+
+export { default as ModalHeader } from './components/ModalHeader';
+export * from './components/ModalHeader';
+
+export { default as ModalHeaderIcon } from './components/ModalHeaderIcon';
+export * from './components/ModalHeaderIcon';
