+++ packages/react/src/components/overlay/index.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as Overlay } from './components/OverlayScreen';
