+++ packages/react/src/components/side-sheet/components/SideSheetHeader.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import { ModalHeader, DialogHeaderProps } from '../../dialog';
+import { ModalHeader, ModalHeaderProps } from '../../modal';
 
-export type SideSheetHeaderProps = DialogHeaderProps;
+export type SideSheetHeaderProps = ModalHeaderProps;
 
 const SideSheetHeader: React.FC<SideSheetHeaderProps> = ModalHeader;
